:truck: refactor: rename modules
