+++ README.md
@@ -105,7 +105,7 @@ python -m bilili.bilili_dl <url>
 |支持程度|中（少数视频不支持）|高|低（仅支持 acg_video）|
 |下载速度|高|低|中|
 |需要 FFmpeg 合并|是|是|否|
-|清晰度支持|高|中（部分较新的 4K 等清晰度无法获取）|少（仅支持 1080P 及更低的清晰度，且无法选择）|
+|清晰度支持|全面|中（部分较新的 4K 等清晰度无法获取）|极少（仅支持 1080P 及更低的清晰度，且无法选择）|
 |总结|B 站当前使用的格式，拥有齐全的清晰度和最佳的下载速度|当 M4S 无法下载时的备用选项，但大多数视频也在支持|除了不需要合并，一无是处|
 
 ### 高清晰度下载
