+++ bilili/api/acg_video.py
@@ -4,7 +4,7 @@ import os
 from bilili.tools import (spider, regex_acg_video_av, regex_acg_video_av_short,
                             regex_acg_video_bv, regex_acg_video_bv_short)
 from bilili.video import BililiContainer
-from bilili.common.base import repair_filename, touch_dir
+from bilili.utils.base import repair_filename, touch_dir
 
 info_api = "https://api.bilibili.com/x/player/pagelist?aid={avid}&bvid={bvid}&jsonp=jsonp"
 parse_api = "https://api.bilibili.com/x/player/playurl?avid={avid}&cid={cid}&bvid={bvid}&qn={qn}&type=&otype=json"
