+++ bilili/api/bangumi.py
@@ -3,7 +3,7 @@ import os
 
 from bilili.tools import spider
 from bilili.video import BililiContainer
-from bilili.common.base import repair_filename, touch_dir
+from bilili.utils.base import repair_filename, touch_dir
 
 info_api = "https://api.bilibili.com/pgc/web/season/section?season_id={season_id}"
 parse_api = "https://api.bilibili.com/pgc/player/web/playurl?avid={avid}&cid={cid}&qn={qn}&ep_id={ep_id}"
