+++ tests/test_acg_video.py
@@ -1,7 +1,7 @@
 import pytest
 
 from bilili.api.acg_video import get_title, get_context, get_containers, parse_segments
-from bilili.common.base import touch_dir
+from bilili.utils.base import touch_dir
 from bilili.video import BililiContainer
