+++ tests/test_bangumi.py
@@ -1,7 +1,7 @@
 import pytest
 
 from bilili.api.bangumi import get_title, get_context, get_containers, parse_segments
-from bilili.common.base import touch_dir
+from bilili.utils.base import touch_dir
 from bilili.video import BililiContainer
